requirements.txt
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
pdfplumber>=0.11
playwright>=1.48
playwright-stealth>=1.0.6
PyYAML>=6.0
python-dotenv>=1.0
jinja2>=3.1
resend>=2.4
pytest>=8.2
